bookhub/requirements.txt
django==5.2.5
django-rest-framework==0.1.0
beautifulsoup4==4.13.4
requests==2.32.4
fake-useragent==2.2.0
redis==6.4.0
django-cors-headers==4.7.0
python-decouple==3.8
redis==6.4.0